ollama_simple_summary/requirements.txt
ollama
newspaper3k